Refactor: Extract components and hooks for better organization
