+++ src/components/ProgressBar.jsx
@@ -1,0 +1,10 @@
+export function ProgressBar({ progress }) {
+  return (
+    <div className="progress-brutal mb-8">
+      <div 
+        className="progress-fill" 
+        style={{ width: `${progress}%` }}
+      ></div>
+    </div>
+  )
+}
