+++ src/components/index.js
@@ -1,0 +1,7 @@
+export { Header } from './Header'
+export { Timer } from './Timer'
+export { Settings } from './Settings'
+export { Statistics } from './Statistics'
+export { ProgressBar } from './ProgressBar'
+export { TimerControls } from './TimerControls'
+export { PresetButtons } from './PresetButtons'
